activation/registration
added email validation function, but i'm not confident that it works since I copied it from
http://www.totallyphp.co.uk/code/validate_an_email_address_using_regular_expressions.htm
The function used is deprecated so I've left a comment saying we should update it.
I'll put it on the to-do list.

added column to the user table indicating whether or not the user has been activated

modified copiable schema to reflect change setting preadded to already activated

added my user account to the copiable schema

modified account.php so that logging was only possible if the user was activated

implemented activation.php code so that users can activate their account

***WARNING***
I don't know if the email code works and what's written there is definitely not enough

Signed-off-by: Chidozie Nwobilor <chidozie.nwobilor@gmail.com>